SVN Commit #86 - 28 May 2011 by CppMonkey
Wow, can't believe I missed the plugin.yml change required to allow setCommand("chest")

Commented various things and updated it to keep it more in line with the hybrid version
A HUGE amount of changes to formatting included
I hope there are no errors that have sneaked in... Also what happens when a creeper damages the chests?

Also chest properties are never saved. I like this as your not waiting time waiting for it to save. BUT it wasn't being saved at all!